refactor: move constants to the top of default.yaml
Problem: Per feedback from volunteer, default.yaml is a bit
overwhelming for non-programmer.

Solution: Move important configuration constant to the top of the file
so it's easier to find and not as daunting.

Risk:
N/A, should fail immediately if mistake was made while moving those
constants because they are all required.

Testing Done:
Test generating result and it works fine.