feat: append entry from source excel file
Problem: When new intake record are added to the excel spreadsheet,
instead of loading them we just reload the cache.

Solution: When we load from cache, also merge in new record if the
flag is set to true.

Risk:
This call will cause new geocoding request to be made for the newly
merged in request, which is expected. If the data is the same as the
cache, the merge should have no impact and result in no geocode call.

Testing Done:
Tested running the script repeatedly and also with new record and made
sure it works as before.