feat(data_loader.excel): pull out config for normalized addr col name
Problem:
We are hardcoding SITEADDRESS for the normalized address column name,
which is tie specifically to the DINS database. We want to make this
more generalize and configurable.

Solution:
Pull out a configuration parameter for the normalized address output.

Risk:
Mistake in code could invalidate the cache and cause geocoding to be
run again on all entries. Mitigated by testing to make sure it can
still load the cache.

Testing Done:
Re-ran the code and make sure the cache hit. Tried running a few
entries without cache and it works too.